pyproject.toml
[project]
name = "downtime-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Cristian Ticu",email = "cristian.ticu@eggnita.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-cov (>=6.1.0,<7.0.0)",
    "types-boto3[essential] (>=1.37.26,<2.0.0)",
    "boto3[crt] (>=1.37.26,<2.0.0)",
    "aws-lambda-typing (>=2.20.0,<3.0.0)",
    "pydantic (>=2.11.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
